Add option to push to Tronbyt without the addon.
This change adds a flag called `tronbyt` to the device configuration.
When set to true, the integration will delegate the rendering to the
Tronbyt server instead of using the add-on. If only Tronbyts are
configured, the integration will not wait for the add-on to be
started, allowing the user to run the integration without the add-on.